add files for a deep dive
I'm planning to investigate the differences in performance between the zig and c implementation. The main reason behind the difference is the simd and because (a+b)+c!=a+(b+c) for floatingpoint numbers we have to use the -mno-fma flag which limits the compiler to get the exactly same result as the zig version.
(Also if this associativity violation is a concern for you don't use double or float, gnu mp exist (and other stuff Tom math etc.). )